requirements.txt
streamlit
psycopg2-binary
fpdf
pillow
pandas
